Store blood sugar level in mg/dl

+++ app/src/main/java/ru/quect/sweetdiary/models/DiaryEntry.java
@@ -1,17 +1,31 @@
 package ru.quect.sweetdiary.models;
 
 public class DiaryEntry {
-    private double mGlucose;
+    private short mGlucose;
 
-    public DiaryEntry(double glucose) {
+    /**
+     * Constructs a new instance of DiaryEntry with the specified initial
+     * blood sugar level in mg/dl units.
+     * @param glucose the initial blood sugar level
+     */
+    public DiaryEntry(short glucose) {
         mGlucose = glucose;
     }
 
-    public double getGlucose() {
+    /**
+     * Returns blood sugar level in mg/dl units.
+     *
+     * @return blood sugar level
+     */
+    public short getGlucose() {
         return mGlucose;
     }
 
-    public void setGlucose(double glucose) {
+    /**
+     * Set blood sugar level in mg/dl units.
+     * @param glucose blood sugar level
+     */
+    public void setGlucose(short glucose) {
         mGlucose = glucose;
     }
 }
